--- grpc/server_test.go
@@ -1,1 +1,0 @@
-package grpc
